harness-wiki/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "harness-wiki"
version = "0.1.0"
description = "Self-optimizing agent harness with persistent wiki memory"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "typer>=0.12",
    "tomli>=2.0",
    "python-dotenv>=1.0",
    "anyio>=4.0",
]

[project.scripts]
harness = "harness_wiki.cli:app"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
